Unify all specializations
