+++ core/src/Kokkos_Half.hpp
@@ -23,6 +23,7 @@
 
 #include <impl/Kokkos_Half_FloatingPointWrapper.hpp>
 #include <impl/Kokkos_Half_NumericTraits.hpp>
+#include <impl/Kokkos_Half_ReductionIdentity.hpp>
 #include <impl/Kokkos_Half_MathematicalFunctions.hpp>
 
 #ifdef KOKKOS_IMPL_PUBLIC_INCLUDE_NOTDEFINED_HALF
