+++ core/src/SYCL/Kokkos_SYCL_Half_Conversion.hpp
@@ -21,11 +21,8 @@
 
 #include <SYCL/Kokkos_SYCL_Half_Impl_Type.hpp>
 #include <impl/Kokkos_Half_FloatingPointWrapper.hpp>
-#include <Kokkos_ReductionIdentity.hpp>
-#include <impl/Kokkos_Half_NumericTraits.hpp>
 
-namespace Kokkos {
-namespace Experimental {
+namespace Kokkos::Experimental {
 
 /************************** half conversions **********************************/
 KOKKOS_INLINE_FUNCTION
@@ -103,33 +100,8 @@ KOKKOS_INLINE_FUNCTION std::enable_if_t<std::is_same_v<T, unsigned long>, T>
 cast_from_half(half_t val) {
   return static_cast<T>(half_t::impl_type(val));
 }
-}  // namespace Experimental
 
-template <>
-struct reduction_identity<Kokkos::Experimental::half_t> {
-  KOKKOS_FORCEINLINE_FUNCTION constexpr static Kokkos::Experimental::half_t
-  sum() noexcept {
-    return Kokkos::Experimental::half_t::impl_type(0.0F);
-  }
-  KOKKOS_FORCEINLINE_FUNCTION constexpr static Kokkos::Experimental::half_t
-  prod() noexcept {
-    return Kokkos::Experimental::half_t::impl_type(1.0F);
-  }
-  KOKKOS_FORCEINLINE_FUNCTION constexpr static auto max() noexcept {
-    // sycl::half doesn't have constexpr constructors so we return
-    // bit_comparison_type which doesn't have a unitary minus operator.
-    // -inf
-    return Kokkos::Experimental::half_t::bit_comparison_type{
-        0b1'11111'0000000000};
-  }
-  KOKKOS_FORCEINLINE_FUNCTION constexpr static auto min() noexcept {
-    // sycl::half doesn't have constexpr constructors so we return
-    // bit_comparison_type
-    return Kokkos::Experimental::infinity_v<Kokkos::Experimental::half_t>;
-  }
-};
-
-}  // namespace Kokkos
+}  // namespace Kokkos::Experimental
 #endif  // KOKKOS_IMPL_SYCL_HALF_TYPE_DEFINED
 
 #ifdef KOKKOS_IMPL_SYCL_BHALF_TYPE_DEFINED
@@ -215,15 +187,6 @@ cast_from_bhalf(bhalf_t val) {
   return static_cast<T>(bhalf_t::impl_type(val));
 }
 }  // namespace Kokkos::Experimental
-
-template <>
-struct Kokkos::reduction_identity<Kokkos::Experimental::bhalf_t> {
-  static constexpr auto inf = Experimental::infinity_v<float>;
-  KOKKOS_FUNCTION static Experimental::bhalf_t sum() { return 0; }
-  KOKKOS_FUNCTION static Experimental::bhalf_t prod() { return 1; }
-  KOKKOS_FUNCTION static Experimental::bhalf_t max() { return -inf; }
-  KOKKOS_FUNCTION static Experimental::bhalf_t min() { return inf; }
-};
 
 #endif  // KOKKOS_IMPL_SYCL_BHALF_TYPE_DEFINED
